WIP add interface to output formatters, rewire up
