+++ packages/ado-extension/src/ioc/setup-ioc-container.ts
@@ -10,6 +10,8 @@ import { AdoIocTypes } from './ado-ioc-types';
 import { ADOArtifactsInfoProvider } from '../ado-artifacts-info-provider';
 import { WorkflowEnforcer } from '../progress-reporter/enforcement/workflow-enforcer';
 import { AdoConsoleCommentCreator } from '../progress-reporter/console/ado-console-comment-creator';
+import { ResultOutputBuilder } from '@accessibility-insights-action/shared/dist/output/result-output-builder';
+import { OutputFormatter } from '@accessibility-insights-action/shared/dist/output/output-formatter';
 
 export function setupIocContainer(container = new inversify.Container({ autoBindInjectable: true })): inversify.Container {
     container = setupSharedIocContainer(container);
@@ -25,6 +27,16 @@ export function setupIocContainer(container = new inversify.Container({ autoBind
         })
         .inSingletonScope();
     container.bind(iocTypes.ArtifactsInfoProvider).to(ADOArtifactsInfoProvider).inSingletonScope();
+    container.bind(ResultOutputBuilder).toSelf().inSingletonScope();
+    container
+        .bind<inversify.interfaces.Factory<ResultOutputBuilder>>('Factory<ResultOutputBuilder>')
+        .toFactory<ResultOutputBuilder, [OutputFormatter]>((context) => {
+            const resultOutputBuilder = context.container.get(ResultOutputBuilder);
+            return (outputFormatter: OutputFormatter): ResultOutputBuilder => {
+                resultOutputBuilder.setOutputFormatter(outputFormatter);
+                return resultOutputBuilder;
+            };
+        });
 
     return container;
 }
