+++ packages/shared/src/index.ts
@@ -7,8 +7,6 @@ export { Scanner } from './scanner/scanner';
 export { iocTypes } from './ioc/ioc-types';
 export { ProgressReporter } from './progress-reporter/progress-reporter';
 export { ReportMarkdownConvertor } from './mark-down/report-markdown-convertor';
-export { productTitle } from './mark-down/markdown-formatter';
-export { disclaimerText } from './content/mark-down-strings';
 export { checkRunDetailsTitle, checkRunName } from './content/strings';
 export { TaskConfig } from './task-config';
 export { BaselineInfo } from './baseline-info';
